Session connection optimization (#4)
In the previous implementation, for SELECT we would create a
new connection each time and then query the relevant content.
For DELETE or UPDATE, we may create multiple connections to
query and update deleted data, which is not good or right.
In this commit, we re-implemented the logic and now our session
connections are reusable. At the same time, each SQL statement
executed will be In a whole transaction, this was not possible before.
The output of some debugging information is also adjusted